Tests for parsing of member names into hosts. Caught a few bugs.
